size

+++ 0x00-hello_world/6-size.c
@@ -5,10 +5,10 @@
  */
 int main(void)
 {
-printf("size of char: %zu bytes(s)\n", (unsigned long)sizeof(char));
-printf("size of int: %zu bytes(s)\n", (unsigned long)sizeof(int));
-printf("size of long int: %zu bytes(s)\n", (unsigned long)sizeof(long int));
-printf("size of long long int: %zu bytes(s)\n", (unsigned long)sizeof(long long int));
-printf("size of float: %zu bytes(s)\n", (unsigned long)sizeof(float));
+printf("size of char: %zu bytes(s)\n", sizeof(char));
+printf("size of int: %zu bytes(s)\n", sizeof(int));
+printf("size of long int: %zu bytes(s)\n", sizeof(long int));
+printf("size of long long int: %zu bytes(s)\n", sizeof(long long int));
+printf("size of float: %zu bytes(s)\n", sizeof(float));
 return (0);
 }
